Fix hassio mocking in ESPHome dashboard tests
I have tried to fix the test_restore_dashboard_storage_skipped_if_addon_uninstalled
test a few times and it never fails locally since the problem depends on order
of test execution. The last failure showed hassio was not being mocked out
properly so for the 3rd attempt to fix it, make sure hassio is mocked out

https://app.codecov.io/gh/home-assistant/core/tests/dev?parameter=FLAKY_TESTS
